+++ src/ADO.Net.Client.Core/IDbObjectFactory.cs
@@ -25,13 +25,12 @@ SOFTWARE.*/
 using System.Collections.Generic;
 using System.Data;
 using System.Data.Common;
-using System.Reflection;
 #endregion
 
 namespace ADO.Net.Client.Core
 {
     /// <summary>
-    /// The contract class for a factory class that creates database objects in order to query a database
+    /// The contract for a factory class that creates database objects in order to query a database
     /// </summary>
     public interface IDbObjectFactory
     {
@@ -97,7 +96,7 @@ namespace ADO.Net.Client.Core
         /// <param name="commandTimeout">Gets or sets the wait time in seconds before terminating the attempt to execute a command and generating an error.</param>
         /// <param name="connection">Represents a connection to a database</param>
         /// <param name="transact">An instance of <see cref="DbTransaction"/></param>
-        /// <param name="parameters">The <see cref="IEnumerable{DbParameter}"/> of parameters associated with the query parameter</param>
+        /// <param name="parameters">The <see cref="IEnumerable{T}"/> of <see cref="DbParameter"/> associated with the query parameter</param>
         /// <param name="query">The SQL command text or name of stored procedure to execute against the data store</param>
         /// <param name="queryCommandType">Represents how a command should be interpreted by the data provider</param>
         /// <returns>Returns an instantiated formatted <see cref="DbCommand"/> object based off the provider passed into the class</returns>
@@ -108,62 +107,10 @@ namespace ADO.Net.Client.Core
         /// <returns>Returns a new instance of the <see cref="DbConnection"/> subclass based on the specified provider</returns>
         DbConnection GetDbConnection();
         /// <summary>
-        /// Gets an <see cref="IEnumerable{T}"/> of <see cref="DbParameter"/> from the passed in <paramref name="values"/>
-        /// </summary>
-        /// <param name="values">An array of values to be used to create <see cref="DbParameter"/></param>
-        /// <returns>Returns an <see cref="IEnumerable{T}"/> of <see cref="DbParameter"/></returns>
-        IEnumerable<DbParameter> GetDbParameters(params object[] values);
-        /// <summary>
         /// Create an instance of a <see cref="DbParameter"/> based off of the provider passed into factory
         /// </summary>
         /// <returns>Returns an instantiated <see cref="DbParameter"/> object</returns>
         DbParameter GetDbParameter();
-        /// <summary>
-        /// Gets an initialized instance of a <see cref="DbParameter"/> subclass based on the specified provider
-        /// </summary>
-        /// <param name="parameterName">The name of the parameter to identify the parameter</param>
-        /// <param name="parameterValue">The value of the parameter</param>
-        /// <returns>Returns an instance of <see cref="DbParameter"/> type with information passed into procedure</returns>
-        DbParameter GetDbParameter(string parameterName, object parameterValue);
-        /// <summary>
-        /// Gets an initialized instance of a <see cref="DbParameter"/>
-        /// </summary>
-        /// <param name="parameterValue">The value t0 assign to the <see cref="DbParameter"/></param>
-        /// <param name="info">An instance of <see cref="PropertyInfo"/></param>
-        /// <returns></returns>
-        DbParameter GetDbParameter(object parameterValue, PropertyInfo info);
-        /// <summary>
-        /// Gets an initialized instance of a <see cref="DbParameter"/> object based on the specified provider
-        /// </summary>
-        /// <param name="dataType">The <see cref="DbType"/> of the field in the database</param>
-        /// <param name="parameterName">The name of the parameter to identify the parameter</param>
-        /// <param name="parameterValue">The value of the parameter as a <see cref="object"/></param>
-        /// <param name="paramDirection">The direction of the parameter, defaults to <see cref="ParameterDirection.Input"/></param>
-        /// <returns>Returns an instance of <see cref="DbParameter"/> object with information passed into procedure</returns>
-        DbParameter GetDbParameter(string parameterName, object parameterValue, DbType dataType, ParameterDirection paramDirection = ParameterDirection.Input);
-#if !NET45
-        /// <summary>
-        /// Gets an initialized instance of a <see cref="DbParameter"/> object based on the specified provider
-        /// </summary>
-        /// <param name="dataType">The <see cref="DbType"/> of the field in the database</param>
-        /// <param name="parameterName">The name of the parameter to identify the parameter</param>
-        /// <param name="parameterValue">The value of the parameter as a <see cref="object"/></param>
-        /// <param name="scale">The number of decimal places to which the <see cref="DbParameter.Value"/> property is resolved.  The default value is <c>null</c></param>
-        /// <param name="precision">The maximum number of digits used to represent the <see cref="DbParameter.Value"/> property.  The default value is <c>null</c></param>
-        /// <param name="paramDirection">The direction of the parameter, defaults to <see cref="ParameterDirection.Input"/></param>
-        /// <returns>Returns an instance of <see cref="DbParameter"/> object with information passed into procedure</returns>
-        DbParameter GetFixedSizeDbParameter(string parameterName, object parameterValue, DbType dataType, byte? scale = null, byte? precision = null, ParameterDirection paramDirection = ParameterDirection.Input);
-#endif
-        /// <summary>
-        /// Gets an initialized instance of a <see cref="DbParameter"/> object based on the specified provider
-        /// </summary>
-        /// <param name="dataType">The <see cref="DbType"/> of the field in the database</param>
-        /// <param name="size">maximum size, in bytes, of the data.  Should not be set for numeric types.  The default value is <c>null</c></param>
-        /// <param name="parameterName">The name of the parameter to identify the parameter</param>
-        /// <param name="parameterValue">The value of the parameter as a <see cref="object"/></param>
-        /// <param name="paramDirection">The direction of the parameter, defaults to <see cref="ParameterDirection.Input"/></param>
-        /// <returns>Returns an instance of <see cref="DbParameter"/> object with information passed into procedure</returns>
-        DbParameter GetVariableSizeDbParameter(string parameterName, object parameterValue, DbType dataType, int? size = null, ParameterDirection paramDirection = ParameterDirection.Input);
         #endregion
     }
 }